requirements.txt
streamlit
openai
speechrecognition
pyttsx3
reportlab
python-dotenv
streamlit_webrtc